# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "次元链接 API",
    "description": "Dimension Link 微服务接口。成功响应为 `{ code: 0, message: 'ok', data }`；失败时 HTTP 状态码按语义返回，body 同样包一层。鉴权头：`Authorization: Bearer {accessToken}`。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "当前服务"
    },
    {
      "url": "http://127.0.0.1:8080",
      "description": "本地开发"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录或注册返回的 accessToken"
      }
    },
    "parameters": {
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "上一页返回的 nextCursor，首页不传"
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 20,
          "minimum": 1,
          "maximum": 50
        },
        "description": "每页条数，默认 20，最大 50"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 1006
          },
          "message": {
            "type": "string",
            "example": "请先登录"
          },
          "data": {
            "nullable": true,
            "example": null
          }
        }
      },
      "UserPublic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "u_me"
          },
          "nickname": {
            "type": "string",
            "example": "星野铃"
          },
          "handle": {
            "type": "string",
            "example": "@hoshi_suzu"
          },
          "bio": {
            "type": "string"
          },
          "signature": {
            "type": "string"
          },
          "emoji": {
            "type": "string",
            "example": "🎀"
          },
          "accentIndex": {
            "type": "integer",
            "minimum": 0,
            "maximum": 7
          },
          "followers": {
            "type": "integer"
          },
          "following": {
            "type": "integer"
          },
          "level": {
            "type": "integer"
          },
          "badges": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isFollowing": {
            "type": "boolean"
          },
          "joinedCircleIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "仅 GET /v1/me 返回"
          }
        }
      },
...
      "CircleSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "c_cos"
          },
          "name": {
            "type": "string",
            "example": "COSPLAY"
          },
          "emoji": {
            "type": "string",
            "example": "👗"
          }
        }
      },
...
      "CommentItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserPublic"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PostCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "p1"
          },
          "author": {
            "$ref": "#/components/schemas/UserPublic"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "mood": {
            "type": "string",
            "enum": [
              "happy",
              "excited",
              "sleepy",
              "love",
              "sad",
              "fire"
            ]
          },
          "circle": {
            "$ref": "#/components/schemas/CircleSummary"
          },
          "imageHue": {
            "type": "integer",
            "minimum": 0,
            "maximum": 359
          },
          "imageTitle": {
            "type": "string"
          },
          "likeCount": {
            "type": "integer"
          },
          "starCount": {
            "type": "integer"
          },
          "commentCount": {
            "type": "integer"
          },
          "liked": {
            "type": "boolean"
          },
          "starred": {
            "type": "boolean"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentItem"
            },
            "description": "仅详情接口可能附带前 N 条"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/me": {
      "get": {
        "tags": [
          "住民"
        ],
        "summary": "当前住民",
        "responses": {
          "200": {
            "description": "含 joinedCircleIds",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserPublic"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "请先登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1006,
                  "message": "请先登录",
                  "data": null
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "住民"
        ],
        "summary": "更新资料",
        "description": "只更新传入字段。handle、emoji、level、badges 本期只读。",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string"
                  },
                  "bio": {
                    "type": "string"
                  },
                  "signature": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新后的当前用户",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserPublic"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "请先登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1006,
                  "message": "请先登录",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/{userId}": {
      "get": {
        "tags": [
          "住民"
        ],
        "summary": "住民主页",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "u_sakurai"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "含 isFollowing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserPublic"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "住民不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1010,
                  "message": "住民不存在",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/{userId}/follow": {
      "post": {
        "tags": [
          "住民"
        ],
        "summary": "关注 / 取关",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "toggle 后的状态",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "isFollowing": {
                          "type": "boolean"
                        },
                        "followers": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "不能关注自己",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1012,
                  "message": "不能关注自己",
                  "data": null
                }
              }
            }
          },
          "404": {
            "description": "住民不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1010,
                  "message": "住民不存在",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts": {
      "get": {
        "tags": [
          "动态"
        ],
        "summary": "动态列表",
        "description": "都不传为广场时间线；authorId 为某人动态；circleId 为圈子动态。倒序。",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "name": "authorId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "circleId",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "c_art"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页动态",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PostCard"
                          }
                        },
                        "nextCursor": {
                          "type": "string",
                          "nullable": true
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "动态"
        ],
        "summary": "发布动态",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content",
                  "circleId"
                ],
                "properties": {
                  "content": {
                    "type": "string",
                    "example": "今日速写：把发卡画成了小行星环。"
                  },
                  "circleId": {
                    "type": "string",
                    "example": "c_art"
                  },
                  "mood": {
                    "type": "string",
                    "enum": [
                      "happy",
                      "excited",
                      "sleepy",
                      "love",
                      "sad",
                      "fire"
                    ],
                    "example": "happy"
                  },
                  "imageTitle": {
                    "type": "string",
                    "example": "行星发卡",
                    "description": "空或省略则为「今日速记」"
                  },
                  "imageHue": {
                    "type": "integer",
                    "example": 312,
                    "description": "省略则随机 0–359"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "完整 PostCard",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PostCard"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "圈子不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1008,
                  "message": "圈子不存在",
                  "data": null
                }
              }
            }
          },
          "422": {
            "description": "先写点什么再发布吧",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1011,
                  "message": "先写点什么再发布吧",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{postId}": {
      "get": {
        "tags": [
          "动态"
        ],
        "summary": "动态详情",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "p1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "可附带 comments 前 N 条",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PostCard"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "动态不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1007,
                  "message": "动态不存在",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{postId}/like": {
      "post": {
        "tags": [
          "动态"
        ],
        "summary": "点赞 toggle",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "以响应布尔值为准",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "liked": {
                          "type": "boolean"
                        },
                        "likeCount": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{postId}/star": {
      "post": {
        "tags": [
          "动态"
        ],
        "summary": "星标 toggle",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "星标他人动态时给作者写通知",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "starred": {
                          "type": "boolean"
                        },
                        "starCount": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{postId}/comments": {
      "get": {
        "tags": [
          "动态"
        ],
        "summary": "评论列表",
        "description": "时间正序",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "分页评论",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CommentItem"
                          }
                        },
                        "nextCursor": {
                          "type": "string",
                          "nullable": true
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "动态"
        ],
        "summary": "发表评论",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string",
                    "example": "星星眼睛好会！"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "新建评论",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CommentItem"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "先写点什么再发布吧",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1011,
                  "message": "先写点什么再发布吧",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
...
  }
}